Documented the hash module
