+++ .github/CODEOWNERS
@@ -1,0 +1,5 @@
+# Default owner
+* @0xCCF4
+
+# Github workflows
+/.github/ @0xCCF4
