+++ src/utils.rs
@@ -1,7 +1,6 @@
-use std::path::{Path, PathBuf};
+use std::path::{PathBuf};
 use std::time::{SystemTime, UNIX_EPOCH};
 use anyhow::{anyhow, Result};
-use crate::data::{File, GeneralHash};
 
 pub trait LexicalAbsolute {
     fn to_lexical_absolute(&self) -> std::io::Result<PathBuf>;
@@ -9,6 +8,7 @@ pub trait LexicalAbsolute {
 
 impl LexicalAbsolute for PathBuf {
     fn to_lexical_absolute(&self) -> std::io::Result<PathBuf> {
+        // https://internals.rust-lang.org/t/path-to-lexical-absolute/14940
         let mut absolute = if self.is_absolute() {
             PathBuf::new()
         } else {
@@ -23,52 +23,6 @@ impl LexicalAbsolute for PathBuf {
         }
         Ok(absolute)
     }
-}
-
-pub fn hash_file<T>(mut reader: T, hash: &mut GeneralHash) -> Result<u64>
-where T: std::io::Read {
-
-    let mut hasher = hash.hasher();
-    let mut buffer = [0; 4096];
-    let mut content_size = 0;
-
-    loop {
-        let bytes_read = reader.read(&mut buffer)?;
-        content_size += bytes_read as u64;
-        if bytes_read == 0 {
-            break;
-        }
-        hasher.update(&buffer[..bytes_read]);
-    }
-
-    *hash = hasher.finalize();
-
-    Ok(content_size)
-}
-
-pub fn hash_directory<'a>(children: impl Iterator<Item = &'a File>, hash: &mut GeneralHash) -> Result<u64> {
-    let mut hasher = hash.hasher();
-
-    let mut content_size = 0;
-
-    for child in children {
-        content_size += 1;
-        hasher.update(child.get_content_hash().as_bytes());
-    }
-
-    *hash = hasher.finalize();
-
-    Ok(content_size)
-}
-
-pub fn hash_path(path: &Path, hash: &mut GeneralHash) -> Result<()> {
-    let mut hasher = hash.hasher();
-
-    hasher.update(path.as_os_str().as_encoded_bytes());
-
-    *hash = hasher.finalize();
-
-    Ok(())
 }
 
 pub fn decode_hex(s: &str) -> Result<Vec<u8>> {
